tracing: Clean up tb_fmt to not give faulty compile warning
gcc incorrectly states that the variable "fmt" is uninitialized when
CC_OPITMIZE_FOR_SIZE is set.

Instead of just blindly setting fmt to NULL, the code is cleaned up
a little to be a bit easier for humans to follow, as well as gcc
to know the variables are initialized.

Change-Id: I7fd0b47fb45d9a5d18ed33d68eae0e754b6a3e61
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>